+++ src/types/product.ts
@@ -1,0 +1,16 @@
+export interface Product {
+  id: number
+  name: string
+  imageURL: string
+  price: {
+    basicPrice: number
+    discountRate: number
+    sellingPrice: number
+  }
+  brandInfo: {
+    id: number
+    name: string
+    imageURL: string
+  }
+  rank: number
+}
